+++ RxCocoa/Traits/Signal/Signal.swift
@@ -33,13 +33,13 @@ public struct SignalSharingStrategy: SharingStrategyProtocol {
     public static var scheduler: SchedulerType { SharingScheduler.make() }
     
     public static func share<Element>(_ source: Observable<Element>) -> Observable<Element> {
-        return source.share(scope: .whileConnected)
+        source.share(scope: .whileConnected)
     }
 }
 
 extension SharedSequenceConvertibleType where SharingStrategy == SignalSharingStrategy {
     /// Adds `asPublisher` to `SharingSequence` with `PublishSharingStrategy`.
     public func asSignal() -> Signal<Element> {
-        return self.asSharedSequence()
+        self.asSharedSequence()
     }
 }