+++ RxCocoa/iOS/UIButton+Rx.swift
@@ -15,7 +15,7 @@ extension Reactive where Base: UIButton {
     
     /// Reactive wrapper for `TouchUpInside` control event.
     public var tap: ControlEvent<Void> {
-        return controlEvent(.touchUpInside)
+        controlEvent(.touchUpInside)
     }
 }
 
@@ -30,7 +30,7 @@ extension Reactive where Base: UIButton {
 
     /// Reactive wrapper for `PrimaryActionTriggered` control event.
     public var primaryAction: ControlEvent<Void> {
-        return controlEvent(.primaryActionTriggered)
+        controlEvent(.primaryActionTriggered)
     }
 
 }
@@ -43,7 +43,6 @@ import RxSwift
 import UIKit
 
 extension Reactive where Base: UIButton {
-    
     /// Reactive wrapper for `setTitle(_:for:)`
     public func title(for controlState: UIControl.State = []) -> Binder<String?> {
         return Binder(self.base) { button, title -> Void in
@@ -74,13 +73,11 @@ extension Reactive where Base: UIButton {
     import UIKit
     
     extension Reactive where Base: UIButton {
-        
         /// Reactive wrapper for `setAttributedTitle(_:controlState:)`
         public func attributedTitle(for controlState: UIControl.State = []) -> Binder<NSAttributedString?> {
             return Binder(self.base) { button, attributedTitle -> Void in
                 button.setAttributedTitle(attributedTitle, for: controlState)
             }
         }
-        
     }
 #endif