+++ RxCocoa/iOS/UIControl+Rx.swift
@@ -12,21 +12,6 @@ import RxSwift
 import UIKit
 
 extension Reactive where Base: UIControl {
-    
-    /// Bindable sink for `enabled` property.
-    public var isEnabled: Binder<Bool> {
-        return Binder(self.base) { control, value in
-            control.isEnabled = value
-        }
-    }
-
-    /// Bindable sink for `selected` property.
-    public var isSelected: Binder<Bool> {
-        return Binder(self.base) { control, selected in
-            control.isSelected = selected
-        }
-    }
-
     /// Reactive wrapper for target action pattern.
     ///
     /// - parameter controlEvents: Filter for observed event types.