+++ RxSwift/Observables/SwitchIfEmpty.swift
@@ -16,7 +16,7 @@ extension ObservableType {
      - returns: Observable sequence that contains elements from switchTo sequence if source is empty, otherwise returns source sequence elements.
      */
     public func ifEmpty(switchTo other: Observable<Element>) -> Observable<Element> {
-        return SwitchIfEmpty(source: self.asObservable(), ifEmpty: other)
+        SwitchIfEmpty(source: self.asObservable(), ifEmpty: other)
     }
 }