+++ RxTest/Recorded.swift
@@ -28,12 +28,12 @@ public struct Recorded<Value>
 extension Recorded {
     /// A textual representation of `self`, suitable for debugging.
     public var debugDescription: String {
-        return "\(self.value) @ \(self.time)"
+        "\(self.value) @ \(self.time)"
     }
 }
 
 extension Recorded: Equatable where Value: Equatable {
     public static func == (lhs: Recorded<Value>, rhs: Recorded<Value>) -> Bool {
-        return lhs.time == rhs.time && lhs.value == rhs.value
+        lhs.time == rhs.time && lhs.value == rhs.value
     }
 }