+++ Tests/RxCocoaTests/KVOObservableTests.swift
@@ -1585,7 +1585,7 @@ extension KVOObservableTests {
 
 extension NSString {
     func duplicate() -> NSString {
-        return NSMutableString(string: self)
+        NSMutableString(string: self)
     }
 }